feat: custom route metric label
Add a the possibility to pass a custom function to generate the route
label, introducing the "static" and "dynamic" mode for route labels generation.
Refactor tests.
Refactor routes metrics configuration.
Introduce a new naming convention for decorators that are not methods, which follows the
pattern:

	metrics*

Add request and reply decorators to send metrics using the route label automatically.

BREAKING CHANGE: the routes metrics options have changed.
The default route prefix is now an empty string and not 'api'.
The dats instance is exported as `metricsCLient`.